maestro_integrador/requirements.txt
python-dotenv
psycopg2
compago
pypi-xmlrpc
